+++ subsys/bluetooth/shell/audio.c
@@ -1760,26 +1760,20 @@ static int cmd_init(const struct shell *sh, size_t argc, char *argv[])
 		bt_audio_capability_register(BT_AUDIO_DIR_SOURCE, &caps_source);
 	}
 
-	if (IS_ENABLED(CONFIG_BT_AUDIO_CAPABILITY)) {
-		if (IS_ENABLED(CONFIG_BT_PAC_SNK_LOC)) {
-			err = bt_audio_capability_set_location(BT_AUDIO_DIR_SINK,
-							       LOCATION);
-			__ASSERT(err == 0, "Failed to set sink location");
+	if (IS_ENABLED(CONFIG_BT_PAC_SNK_LOC)) {
+		err = bt_audio_capability_set_location(BT_AUDIO_DIR_SINK, LOCATION);
+		__ASSERT(err == 0, "Failed to set sink location");
 
-			err = bt_audio_capability_set_available_contexts(BT_AUDIO_DIR_SINK,
-									 CONTEXT);
-			__ASSERT(err == 0, "Failed to set sink available contexts");
-		}
+		err = bt_audio_capability_set_available_contexts(BT_AUDIO_DIR_SINK, CONTEXT);
+		__ASSERT(err == 0, "Failed to set sink available contexts");
+	}
 
-		if (IS_ENABLED(CONFIG_BT_PAC_SRC_LOC)) {
-			err = bt_audio_capability_set_location(BT_AUDIO_DIR_SOURCE,
-							       LOCATION);
-			__ASSERT(err == 0, "Failed to set source location");
+	if (IS_ENABLED(CONFIG_BT_PAC_SRC_LOC)) {
+		err = bt_audio_capability_set_location(BT_AUDIO_DIR_SOURCE, LOCATION);
+		__ASSERT(err == 0, "Failed to set source location");
 
-			err = bt_audio_capability_set_available_contexts(BT_AUDIO_DIR_SOURCE,
-									 CONTEXT);
-			__ASSERT(err == 0, "Failed to set source available contexts");
-		}
+		err = bt_audio_capability_set_available_contexts(BT_AUDIO_DIR_SOURCE, CONTEXT);
+		__ASSERT(err == 0, "Failed to set source available contexts");
 	}
 
 #if defined(CONFIG_BT_AUDIO_UNICAST)
